requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
PyJWT>=2.8.0
python-dateutil>=2.8.0
